fix / strip reply prefix on history
